Ticket client
This is the SQL script that will add a new row to the target table. If the transaction commits it will add a new ticket to the stack, or it will put back the ticket it has acquired by rolling back the transaction.

This script should run from a client application. For a simple multithread C# sample check this code: https://github.com/ArmandoLacerda/TicketsWithNoGap/blob/master/program.cs .